furniture Domain 변경 LogBack 설정

+++ handicraft-web/handicraft-core/src/main/java/com/handicraft/core/dto/FurnitureToMarket.java
@@ -26,7 +26,7 @@ public class FurnitureToMarket implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "m_id")
-    protected int mid;
+    protected long mid;
 
     protected boolean admin;
